Sladjan1 Updated class name
